Working with easyOCR
